Update index.md
When trying to use the restore_mode on the switch, the behavior of the binary_sensor's button causes the restore_mode state to be respected when a device starts without power, but when an OTA was applied it would cause the switch to do the opposite of what the restore_mode was defined for that switch (i.e. if Relay1 was set to ALWAYS_OFF, the OTA would cause the relay to turn on during reboot, but when disconnecting power (cold boot) it would respect the ALWAYS_OFF mode.

The updates to this file now make it so it behaves as expected - whether an OTA is applied or if the power is disconnected - the restore_mode will be respected.